Fixes of rebase

+++ src/main/java/wolox/training/repositories/UsersRepository.java
@@ -3,6 +3,6 @@ package wolox.training.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import wolox.training.models.Users;
 
-public interface UsersRepository extends JpaRepository<Users, Integer> {
+public interface UsersRepository extends JpaRepository<Users, Long> {
   Users findByUsername(String username);
 }
